requirements.txt
googleapis-common-protos==1.65.0
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
multidict==6.1.0
pillow==10.4.0
protobuf==5.28.0
typing_extensions==4.12.2
viam-sdk==0.29.1
